+++ .travis-ci.sh
@@ -1,8 +1,6 @@
-bash -ex .travis-opam.sh
-
-eval `opam config env`
-opam install mirage -y
+opam depext -uiy mirage
 git clone -b mirage-dev https://github.com/mirage/mirage-skeleton.git
 cd mirage-skeleton
+eval `opam config env`
 MODE=unix make
 MODE=xen  make